Corregido API Products, Suppliers Entero y Orders Gets

+++ ERP_2021_2022_Grupo_1/ERP_2021_2022_Grupo_1_Entities/clsOrder.cs
@@ -8,14 +8,14 @@ namespace ERP_2021_2022_Grupo_1_Entities
     {
         #region attributes
         private int orderId;
-        private int total;
+        private double total;
         private DateTime orderDate;
         private DateTime limitOrderDate;
         private string notes;
         private int supplierId;
         #endregion
         #region constructor
-        public clsOrder(int orderId, int total, DateTime orderDate, DateTime limitOrderDate, string notes, int supplierId)
+        public clsOrder(int orderId, double total, DateTime orderDate, DateTime limitOrderDate, string notes, int supplierId)
         {
             this.orderId = orderId;
             this.total = total;
@@ -36,7 +36,7 @@ namespace ERP_2021_2022_Grupo_1_Entities
         #endregion
         #region public methods
         public int OrderId { get => orderId; }
-        public int Total { get => total; }
+        public double Total { get => total; }
         public DateTime OrderDate { get => orderDate;  }
         public DateTime LimitOrderDate { get => limitOrderDate; }
         public string Notes { get => notes; set => notes = value; }
